DB_fill/requirements.txt
absl-py==0.8.1
asgiref==3.2.3
astor==0.8.0
astunparse==1.6.3
audioread==2.1.8
biopython==1.78
cachetools==3.1.1
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
dill==0.3.1.1
Django==3.0
et-xmlfile==1.0.1
ete3==3.1.2
Flask==1.1.2
future==0.18.2
gast==0.3.3
google-auth==1.7.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
image==1.5.27
importlib-metadata==1.6.0
inflect==4.1.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
joblib==0.14.1
jsonpatch==1.25
jsonpointer==2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.32.1
lxml==4.5.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
multiprocess==0.70.9
mysql-connector-python==8.0.20
numba==0.49.1
numpy==1.18.2
oauthlib==3.1.0
openpyxl==3.0.5
opt-einsum==3.1.0
pafy==0.5.5
pandas==0.25.3
Pillow==6.2.1
pprint==0.1
progress==1.5
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pygame==1.9.6
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
python-dateutil==2.8.1
python-vlc==3.0.7110
pytils==0.3
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.1
qtgui==0.0.1
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.22
scipy==1.4.1
six==1.14.0
sklearn==0.0
sounddevice==0.3.15
SoundFile==0.10.3.post1
sqlparse==0.3.0
tbb==2019.0
tensorboard==1.14.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tools==0.1.9
torch==1.5.0
torchfile==0.1.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.46.0
umap-learn==0.4.3
Unidecode==1.1.1
urllib3==1.25.5
visdom==0.1.8.9
webrtcvad==2.0.10
websocket-client==0.57.0
Werkzeug==0.16.0
wrapt==1.11.2
youtube-dl==2020.3.24
zipp==3.1.0
